:recycle: Separate ResponseError
